usb: dwc2: disable HNP and SRP on stm32mp1 hsotg
With Type-C on stm32mp1, Vbus and ID are not wired to OTG controller.
So Vbus and ID sensing cannot be used (VBDEN and IDDEN in GGPIO register).
However, if HNP and/or SRP support is enabled, Vbus comparators are
automatically enabled independently of VBDEN value.
To avoid bad level detection, which ends in false over-current detection,
HNP and SRP have to be disabled.
Even with Type-µAB, HNP and SRP are not used, because OTG is obsolete.
So disable HNP and SRP on stm32mp1 hsotg.

Fixes: 1b6e92f31749 ("usb: dwc2: add support for STM32MP1 USB OTG HS")

Change-Id: I85a6b3ecc593427b73303903ed14261db9551f2f
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/129930
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>