usb: dwc2: fix wakeup support with dedicated wake irq
In case of -EPROBE_DEFER when trying to get the vbus regulator or enabling
the low level hardware, the wakeirq is not released.
On the next probe, a
genirq: Flags mismatch irq 60. 00002001 (49000000.usb-otg:wakeup) vs. 00002001 (49000000.usb-otg:wakeup)
error is raised, and the probe fails.

This patch cleans error path, to release the wakeirq properly.

Change-Id: Icebe140c14a16831f3f1cebd566b95ac54b72036
Fixes: 28daba17e41e ("usb: dwc2: add wakeup support with dedicated wake irq")
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/172366
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>